Split test_drawdown_recovery.py, fix async antipatterns in test suite
Workstream 1: Split 1,270-line test_drawdown_recovery.py into 3 focused
files by concern, extract 5 duplicated helpers into tests/conftest.py:
- test_drawdown_recovery.py: recovery timer logic (22 tests)
- test_drawdown_starting_equity.py: starting equity + stale state (17 tests)
- test_drawdown_panic_gate.py: _panic_is_live flag gating (10 tests)

Workstream 2: Fix test_thesis_coherence.py where 4 classes inherited
unittest.TestCase with async methods — coroutines were never awaited,
so 28 tests silently passed without running assertions. Changed to
IsolatedAsyncioTestCase and fixed 2 exposed mock gaps (avgFillPrice,
reqAllOpenOrdersAsync).

Workstream 3: Fix test_order_manager.py manual asyncio.run() wrappers
by switching TestOrderManager and TestPositionClosing to
IsolatedAsyncioTestCase.

All 865 tests pass.
